realtime_sync.py
import threading
import time
from datetime import datetime
from classroom_sync import sync_manager
from database import log_sync_activity

class RealtimeSyncService:
    def __init__(self):
        self.sync_interval = 300  # 5 minutes
        self.is_running = False
        self.thread = None
    
    def start_sync_service(self):
        """Start the real-time sync service"""
        self.is_running = True
        self.thread = threading.Thread(target=self._sync_loop)
        self.thread.daemon = True
        self.thread.start()
        print("Real-time Google Classroom sync service started")
    
    def stop_sync_service(self):
        """Stop the real-time sync service"""
        self.is_running = False
        if self.thread:
            self.thread.join()
        print("Google Classroom sync service stopped")
    
    def _sync_loop(self):
        """Main sync loop"""
        while self.is_running:
            try:
                self._sync_pending_activities()
                self._sync_student_enrollments()
                time.sleep(self.sync_interval)
            except Exception as e:
                print(f"Sync error: {e}")
                time.sleep(60)  # Wait 1 minute on error
    
    def _sync_pending_activities(self):
        """Sync pending activities to Google Classroom"""
        # Implement based on your activity system
        pass
    
    def _sync_student_enrollments(self):
        """Sync new student enrollments"""
        # Implement based on your enrollment system
        pass

# Global sync service instance
realtime_sync = RealtimeSyncService()