requirements.txt
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
Werkzeug==2.3.7
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-dotenv==1.0.0
jotform==1.3.0
requests==2.28.2